+++ CHANGELOG.md
@@ -9,7 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 ### Removed
 
-## [1.0.1] - 16.06.2023
+
+## [1.0.2] - 16.06.2023
+- onCacheTags log the class type generating the tags
+- add StoreApiRouteCacheKeyEvent logging
+
+## [1.0.1] - 15.06.2023
 - add more configuration ways so user can decide what he really wants to log
 
 ## [1.0.0] - 21.05.2023